WE HAVE A GAME
Fixed spin bug, did some work on AI (it isn't good just yet, but
passable), fixed multiple collision bug. Overall, this is a passable
version of the game we could do our report on. Some things to refine:
Mouse cursor doesn't align with paddle (maybe figure out a way to hide
it?)
AI needs to be better.

If you can get those things, we'll be ready for a video soon enough.